Isbndbex v0.0.1 Isbndbex

Functions

get_author(pid, id)

Sends a message to the process pid to get the details of the author with the given id. The id corresponds to the internal isbndb author id.

get_authors(pid, query)

Sends a message to the process pid to search for authors who have names similar to query.

get_book(pid, id)

Sends a message to the process pid to get the book with the given id. The id can be the books’s isbn10, isbn13 or the internal isbndb book id.

get_books(pid, query, index)

Sends a message to the process pid to search the index for books matching query. The query parameter expected value changes depending on which index is consulted.

get_categories(pid, query)

Sends a message to the process pid to search for subjects that have categories similar to query.

get_category(pid, id)

Sends a message to the process pid to get the details of the category with the given id. The id corresponds to the internal isbndb category id.

get_prices(pid, id)

Sends a message to the process pid to get a list of the stores selling the book with the given id. The id can be the book’s isbn10, isbn13 or the internal isbndb id.

get_publisher(pid, id)

Sends a message to the process pid to get the details of the publisher with the given id. The id corresponds to the internal isbndb publisher id.

get_publishers(pid, query)

Sends a message to the process pid to search for publishers that have names similar to query.

get_subject(pid, id)

Sends a message to the process pid to get the details of the subject with the given id. The id corresponds to the internal isbndb subject id.

get_subjects(pid, query)

Sends a message to the process pid to search for subjects that have names similar to query.

start(key)

Starts a process with the given API key.